✨	새 기능 : 문제 받아오는 기반
service로직은 변경점을 크게 가져갈꺼 같아서 나머지만 먼저 저장함

+++ src/main/java/darkoverload/itzip/feature/algorithm/entity/ProblemEntity.java
@@ -28,7 +28,7 @@ public class ProblemEntity {
     private Integer level; // 문제 난이도
 
     @Column(name = "accepted_user_count")
-    private Integer acceptedUserCount; // 문제를 푼 유저 수
+    private Long acceptedUserCount; // 문제를 푼 유저 수
 
     private Integer averageTries; // 평균 시도 횟수
